Update SAI BCM-DNX module to compile for 5.10 kernel (#3)
Changes done here are similar as the ones done for
platform/broadcom/saibcm-modules.

Update the build rules in saibcm-modules to use the 5.10 kernel instead
of searching for the 4.19 kernel. In addition, some code changes were
done to get it to compile. The main categories of such changes are as
follows:

* For /proc files, `struct file_operations` has been replaced with
`struct proc_ops`.
* Y2038 changes to use the new APIs, since `do_gettimeofday()` is no
longer available.
* Minor changes in how external kernel module symbols are read by
modpost.
* Don't build and package the psample module, since this is coming
from the kernel.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>